requirements.txt
flask==3.1.1
firebase-admin==7.1.0
openai==1.99.9
python-dotenv==1.1.1
Pillow==11.1.0
requests==2.32.4